requirements.apiserver.txt
adiscstudies==0.11.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
boto3==1.40.21
botocore==1.40.21
brotli==1.1.0
cattrs==25.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
fastapi==0.116.1
fonttools==4.59.2
h11==0.16.0
idna==3.10
jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.9
llvmlite==0.44.0
markupsafe==3.0.2
matplotlib==3.10.6
networkx==3.5
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==25.2
psycopg==3.2.9
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyjwt==2.10.1
pyparsing==3.2.3
pyshp==2.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
s3transfer==0.13.1
scipy==1.16.1
secure==1.0.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
termplotlib==0.3.9
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
zstandard==0.24.0
